feat: updated example
